pyproject.toml
[project]
name = "empty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google>=3.0.0",
    "google-genai>=1.52.0",
    "laminar>=2022.5a1",
    "lmnr[all]>=0.7.23",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
]
